+++ tests/unit/cli_test.py
@@ -27,7 +27,7 @@ def validator_mock(
     self: object,  # noqa: ARG001
     files: object,  # noqa: ARG001
     checksum_module: Optional[str] = None,  # noqa: ARG001
-    skip_model_verification: bool = False,  # noqa: ARG001
+    cache: Optional[object] = None,  # noqa: ARG001
     skip_message_verification: bool = False,  # noqa: ARG001
     split_disjunctions: bool = False,  # noqa: ARG001
 ) -> None: